backend-project/requirements/base.txt
Django==3.1.2
django-environ==0.4.5
django-teryt-tree==0.16.3
djangorestframework==3.12.1
psycopg2-binary==2.8.6
django-extensions==3.0.9
drf-nested-routers==0.92.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
djangorestframework-filters==1.0.0.dev2
drf-yasg2==1.19.4
minio==6.0.0
pyjwt==1.7.1
requests==2.24.0
requests_oauthlib==1.3.0
whitenoise==5.2.0
django-filter==2.4.0
zipstream-new==1.1.8
# add packaging - hard requirements of drf-yasg2
# see https://github.com/JoelLefkowitz/drf-yasg/pull/44
packaging>=18.0
icalendar==4.0.7
django-fullurl==1.1
Sphinx==3.3.1
sphinx-rtd-theme==0.5.0
whitenoise==5.2.0
